+++ src/components/AppLogo/AppLogo.styled.ts
@@ -1,0 +1,10 @@
+import { Colors } from '../../Colors';
+import styled from 'styled-components';
+
+export const Logo = styled.div`
+  color: ${Colors.PrimaryRed};
+  font-size: 1.1em;
+  :first-child {
+    font-weight: 1000;
+  }
+`;
