+++ src/components/AppLogo/AppLogo.tsx
@@ -1,0 +1,8 @@
+import { Logo } from './AppLogo.styled';
+
+export const AppLogo = () => (
+  <Logo>
+    <span>netflix</span>
+    <span>roulette</span>
+  </Logo>
+);
