+++ src/components/MovieDetailsRoot/MovieDetailsRoot.tsx
@@ -1,7 +1,7 @@
-import { AppLogo } from '../../App';
 import { Icon } from '../../App.styled';
 import { MoviesService } from '../../services/movies.service';
 import { Movie } from '../../types/movie';
+import { AppLogo } from '../AppLogo/AppLogo';
 import { MovieDetails } from '../MovieDetails/MovieDetails';
 import { TopContainerHeader } from '../MovieListPage/MovieListPage.styled';
 import {
@@ -30,11 +30,11 @@ export function MovieDetailsRoot() {
   return (
     <DetailsContainer>
       <TopContainerHeader>
-        <AppLogo />{' '}
+        <AppLogo />
         <Link to={linkUrl}>
           <SearchSwitcherButton>
             <Icon src={SearchLogoIcon} />
-          </SearchSwitcherButton>{' '}
+          </SearchSwitcherButton>
         </Link>
       </TopContainerHeader>
       <MovieDetails movie={movie} />
